+++ jobsdb/jobsdb_test.go
@@ -1190,8 +1190,8 @@ func TestThreadSafeJobStorage(t *testing.T) {
 		require.Equal(t, 2, len(jobsDB2.getDSListSnapshot()), "expected jobsDB2 to have refreshed its ds list")
 
 		require.Equal(t, 1, len(jobsDB3.getDSListSnapshot()), "expected jobsDB3 to still have a list length of 1")
-		errorsMap := jobsDB3.StoreEachBatchRetry(context.Background(), [][]*JobT{generateJobs(2)})
-		require.Equal(t, 0, len(errorsMap))
+		err = jobsDB3.Store(context.Background(), generateJobs(2))
+		require.NoError(t, err)
 
 		require.Equal(t, 2, len(jobsDB3.getDSListSnapshot()), "expected jobsDB3 to have refreshed its ds list")
